Tested part of Gameboard function placeShip
Tested the case of the vertical position of just
one ship, and how the adjacencyList is going to
be.
Need to write the test (may be deleting the test
made in this commit) for more than one ship, and
how to check in the adjacencyList if the place is
already taken.
Then, how to resolve this conflict of placement.