Design fork
Two ways of storing the placement of the ships
info:

The first is in the same adjacencyList, storing
both the position and the ship. This is handy but
makes the function that checks if the place is
already taken more cumbersome.

The second is storing the position in each ship,
and only storing which places are available or not
in the adjacency list. In other words, the
adjacencyList only knows if a place is taken or
not, but doesnt know by which ship. Each ship has
to have the functionality to store their own
placement.

So im going to create a branch to try to implement
the latter.