Finished with the DOM manipulation interactivity
Created the elements object which contains all the
elements of the html file that JS need.

Added the functions to manipulate the DOM:
openModeDialog, cancelDialog, approveDialog and
gameModeSelection. Also createElement, createGrid,
and applyGrid.
There is one function called gameInit, but im
going to change it later.

Currently working on the functions to ineract with
the game functionality, such as gameModeData,
displayBoard, displayShip, displayHit, game, etc.