Making changes in the classes for storage
To decouple the code, i want to store the state of
the game in localStorage. Because i cant store the
functions of the objects, just its data, i have to
instantiate a new object with the old state.
Thats why the need to change the classes
constructors, to allow them to take the stored
object data.