Added two new files
The first is storage.js and it contain two
functions: storeData and retrieveData. Now the
state of the game can be stored in localStorage
and then instantiated in its respective objects.
In order to make this new instantiation possible,
the classes have a new method called setProperties
that update the state of the objects.

The second file is display.js, where are going to
be all the functions that take care of rendering
that state in the interface.

So the dom.js file is going to have the functions
that deal with pop-ups and game mode selection.