Added functionality to display hits and miss
In the display file, the new functions displayHit
and displayMiss renders on the grid if the user
target a ship or not.
Also the displayShip function was redesign, and
the createRandomShips function was created.
displayBoard and displayTotalBoard functions
renders all the state of the game, or just a part.

In the storage file, added the upgradeData
function.