Create getters in the classes
The idea is that the DOM functions only interact
with the Game class, not with Player or Gameboard
classes.
So in order to do that, the Game class implement
functions that call Player functions, that in some
cases call Gameboard functions.
Clearly not the best design but this way if i
want to change something in the future, for
example a function in Player, may be i need to
refactor the Game function that call it and NOT
the DOM function.

Also, now the buttons of the players for placing
random ships are disabled when the game starts.