Add reset and replay functionality
Game class has reset and replay functions.
Both removes all the data of the state of the game
but replay keep the name of the players.
Also added the functions in the DOM to
reinitialize its state, like creating a new board
without classes and allowing the users to click on
the previously unabled buttons.
Finally, created the html dialog to display the
winner and the options to replay or reset.